refactor: update model imports to use model module
